Adopt the packagedef file structure
